+++ src/main/java/Rubix.java
@@ -1,7 +1,8 @@
-import com.google.gson.*;
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
 import com.sun.net.httpserver.HttpServer;
+import handlers.RubixHandler;
+import handlers.RubixResHandler;
+
+
 import java.io.*;
 import java.net.InetSocketAddress;
 
@@ -23,37 +24,11 @@ public class Rubix {
     }
 
     private void HTTPStream() throws IOException {
-        R3xResHandler resHandler = new R3xResHandler(this.handler);
+        RubixResHandler resHandler = new RubixResHandler(this.handler);
         HttpServer server = HttpServer.create(new InetSocketAddress(8080), 0);
         server.createContext("/", resHandler);
         server.setExecutor(null);
         server.start();
-    }
-
-    public class R3xResHandler implements HttpHandler {
-        private RubixHandler handler;
-
-        R3xResHandler(RubixHandler r3x){
-            this.handler = r3x;
-        }
-
-        @Override
-        public void handle(HttpExchange httpExchange) throws IOException {
-            InputStream input = httpExchange.getRequestBody();
-            StringBuilder sb = new StringBuilder();
-            new BufferedReader(new InputStreamReader(input))
-                    .lines()
-                    .forEach((String s) -> sb.append(s));
-            JsonParser parser = new JsonParser();
-            JsonObject json = (JsonObject) parser.parse(sb.toString());
-            String response = String.valueOf(this.handler.execute(json));
-            httpExchange.getResponseHeaders().set("Content-Type", "application/json");
-            httpExchange.sendResponseHeaders(200, response.length());
-            OutputStream os = httpExchange.getResponseBody();
-            os.write(response.getBytes());
-            os.close();
-        }
-
     }
 }
