+++ src/main/java/handlers/JsonHandler.java
@@ -1,4 +1,4 @@
-package util;
+package handlers;
 
 import com.google.gson.*;
