+++ src/main/java/handlers/RubixHandler.java
@@ -1,3 +1,5 @@
+package handlers;
+
 import com.google.gson.JsonObject;
 
 public interface RubixHandler {
